Created separate Avatar component, minor code adjustments regarding readability
