+++ frontend/src/components/features/avatar/Avatar.scss
@@ -1,0 +1,71 @@
+@use '../../../common/styles/variables' as *;
+
+.player-avatar {
+  width: 386px;
+  height: 280px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-evenly;
+  border-radius: 10px;
+  background-color: #ffffff;
+  box-shadow: 0px 4px #c3c1b7;
+
+  h2 {
+    padding: 0px;
+    margin: 0px;
+  }
+
+  &__username {
+    color: $darkerPrimaryColor;
+  }
+
+  &__image {
+    background-color: #1e1e1e;
+  }
+
+  &__score {
+    color: $secondaryColor;
+    font-size: 28px;
+  }
+}
+
+.pos-top-left {
+  grid-column: 1;
+  grid-row: 1;
+}
+
+.pos-top {
+  grid-column: 2;
+  grid-row: 1;
+}
+
+.pos-top-right {
+  grid-column: 3;
+  grid-row: 1;
+}
+
+.pos-right {
+  grid-column: 3;
+  grid-row: 2;
+}
+
+.pos-bottom-right {
+  grid-column: 3;
+  grid-row: 3;
+}
+
+.pos-bottom {
+  grid-column: 2;
+  grid-row: 3;
+}
+
+.pos-bottom-left {
+  grid-column: 1;
+  grid-row: 3;
+}
+
+.pos-left {
+  grid-column: 1;
+  grid-row: 2;
+}
