+++ frontend/src/components/pages/room/RoomPage.tsx
@@ -7,9 +7,7 @@ const MockPlayersArr: Player[] = [
   { name: 'Player 1', score: 0 },
   { name: 'Player 2', score: 0 },
   { name: 'Player 3', score: 0 },
-  { name: 'Player 1', score: 0 },
-  { name: 'Player 2', score: 0 },
-  { name: 'Player 3', score: 0 },
+  { name: 'Player 4', score: 0 },
 ];
 
 export const RoomPage = () => {
